RangeDecompressor: switch from float to double in calculations
Single precision is not sufficient to guarantee stable numerical values in the decoded output to the 8th decimal position output by the ASCII messages.

Also improve the equality asserts to display the mismatching string contents on failure.

+++ include/novatel_edie/decoders/oem/rangecmp/range_decompressor.hpp
@@ -12,7 +12,7 @@
 // |  The above copyright notice and this permission notice shall be included    |
 // |  in all copies or substantial portions of the Software.                     |
 // |                                                                             |
-// |  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR |
+// |  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS xOR |
 // |  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,   |
 // |  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL    |
 // |  THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER |
@@ -107,12 +107,12 @@ class RangeDecompressor
 
   private:
     static double GetSignalWavelength(const ChannelTrackingStatusStruct& stChannelTrackingStatus_, int16_t sGLONASSFrequency_);
-    float DetermineRangeCmp2ObservationLockTime(const MetaDataStruct& stMetaData_, uint32_t uiLockTimeBits_,
-                                                ChannelTrackingStatusStruct::SATELLITE_SYSTEM eSystem_,
-                                                ChannelTrackingStatusStruct::SIGNAL_TYPE eSignal_, uint16_t usPRN_);
-    float DetermineRangeCmp4ObservationLockTime(const MetaDataStruct& stMetaData_, uint8_t ucLockTimeBits_,
-                                                ChannelTrackingStatusStruct::SATELLITE_SYSTEM eSystem_,
-                                                ChannelTrackingStatusStruct::SIGNAL_TYPE eSignal_, uint32_t uiPRN_);
+    double DetermineRangeCmp2ObservationLockTime(const MetaDataStruct& stMetaData_, uint32_t uiLockTimeBits_,
+                                                 ChannelTrackingStatusStruct::SATELLITE_SYSTEM eSystem_,
+                                                 ChannelTrackingStatusStruct::SIGNAL_TYPE eSignal_, uint16_t usPRN_);
+    double DetermineRangeCmp4ObservationLockTime(const MetaDataStruct& stMetaData_, uint8_t ucLockTimeBits_,
+                                                 ChannelTrackingStatusStruct::SATELLITE_SYSTEM eSystem_,
+                                                 ChannelTrackingStatusStruct::SIGNAL_TYPE eSignal_, uint32_t uiPRN_);
     template <bool bIsSecondary>
     void DecompressReferenceBlock(uint8_t** ppucDataPointer_, RangeCmp4MeasurementSignalBlockStruct& stReferenceBlock_,
                                   MEASUREMENT_SOURCE eMeasurementSource_);
